make open_files a model method
